:sparkles: feat(voucherAmount): voucher amount 관련 클래스 구현
