+++ src/main/java/com/dojinyou/devcourse/voucherapplication/voucher/domain/FixedAmount.java
@@ -1,14 +1,13 @@
-package com.dojinyou.devcourse.voucherapplication.voucher.Entity;
+package com.dojinyou.devcourse.voucherapplication.voucher.domain;
 
-public class PercentAmount implements VoucherAmount {
+public class FixedAmount extends VoucherAmount {
     private static final int MIN_AMOUNT = 0;
-    private static final int MAX_AMOUNT = 100;
+    private static final int MAX_AMOUNT = 1_000_000_000;
     public static final String ERROR_MESSAGE_OUT_OT_RANGE = "voucher amount 범위를 벗어났습니다.";
-    private final int amount;
 
-    public PercentAmount(int amount) {
+    public FixedAmount(int amount) {
+        super(amount);
         validate(amount);
-        this.amount = amount;
     }
 
     @Override
@@ -16,9 +15,5 @@ public class PercentAmount implements VoucherAmount {
         if (amount < MIN_AMOUNT || MAX_AMOUNT < amount) {
             throw new IllegalArgumentException(ERROR_MESSAGE_OUT_OT_RANGE);
         }
-    }
-
-    public int getAmount() {
-        return amount;
     }
 }
